+++ src/main/java/com/cleanroommc/ivgtmb/proxy/CommonProxy.java
@@ -1,0 +1,13 @@
+package com.cleanroommc.ivgtmb.proxy;
+
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+public class CommonProxy {
+
+    public void preInit(FMLPreInitializationEvent event) {
+    }
+
+    public void init(FMLInitializationEvent event) {
+    }
+}